requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.2
django-bootstrap4==3.0.0
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
psycopg2==2.8.6
python-decouple==3.4
pytz==2021.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
